requirements.txt
playwright>=1.40.0
img2pdf>=0.5.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0
